refactor: extract metoo button logic to custom hook
