+++ src/pages/DetailPage/components/MetooButton/MetooButton.tsx
@@ -1,54 +1,16 @@
-import { FC, useEffect, useState } from 'react';
+import { FC } from 'react';
 import { styled } from 'styled-components';
 import { Theme } from '../../../../styles';
 import { UncomBigIcon } from '../../../../assets';
-import { useDetailData } from '../../container';
-import { toggleMetooData } from '../../../../api';
 type MetooButtonProps = {
-  agreeCount: number;
-  onClick: () => void;
-  agree: boolean;
-  postId: string;
+  handleClickMetooButton: () => void;
+  localMetooCount: number | null;
 };
 
 export const MetooButton: FC<MetooButtonProps> = ({
-  agreeCount = 0,
-  onClick,
-  agree = false,
-  postId,
+  handleClickMetooButton,
+  localMetooCount,
 }) => {
-  const { isLoading, error } = useDetailData();
-  const [localMetoo, setLocalMetoo] = useState<boolean | null>(null);
-  const [localMetooCount, setLocalMetooCount] = useState<number | null>(null);
-
-  if (isLoading) {
-    return <></>;
-  }
-
-  const handleClickMetooButton = async () => {
-    await toggleMetooData(postId);
-
-    // TODO
-    const token = localStorage.getItem('token');
-
-    if (token) {
-      if (localMetoo) {
-        setLocalMetooCount((prevCount) => (prevCount ? prevCount - 1 : 0));
-        setLocalMetoo(false);
-      } else {
-        setLocalMetooCount((prevCount) => (prevCount ? prevCount + 1 : 1));
-        setLocalMetoo(true);
-      }
-    } else {
-      alert('로그인 후 이용 가능합니다');
-    }
-  };
-
-  useEffect(() => {
-    setLocalMetoo(agree);
-    setLocalMetooCount(agreeCount);
-  }, [agree, agreeCount]);
-
   return (
     <StButton onClick={handleClickMetooButton}>
       <StButtonLabel>
