+++ src/pages/DetailPage/container/hooks/detailData.ts
@@ -19,7 +19,34 @@ export const useDetailData = () => {
     };
   }
 
-  const { owner } = data;
+  const {
+    owner,
+    agree,
+    agreeCount,
+    imageUrlList,
+    title,
+    content,
+    createdAt,
+    address,
+    nickname,
+    done,
+    doneCount,
+  } = data;
 
-  return { data, isLoading, error, postId, owner };
+  return {
+    isLoading,
+    error,
+    postId,
+    owner,
+    agree,
+    agreeCount,
+    imageUrlList,
+    title,
+    content,
+    createdAt,
+    address,
+    nickname,
+    done,
+    doneCount,
+  };
 };
