+++ src/pages/DetailPage/views/UnderButtons/UnderButtonsView.tsx
@@ -4,27 +4,23 @@ import { LocationButton, MetooButton } from '../../components';
 type UnderButtonsViewProps = {
   address: string;
   owner: boolean;
-  agree: boolean;
-  agreeCount: number;
-  postId: string;
+  handleClickMetooButton: () => void;
+  localMetooCount: number | null;
 };
 
 export const UnderButtonsView: FC<UnderButtonsViewProps> = ({
   address,
   owner = false,
-  agree,
-  agreeCount,
-  postId,
+  handleClickMetooButton,
+  localMetooCount,
 }) => {
   return (
     <StWrapper>
       <StButtonWrapper>
         <LocationButton address={address} />
         <MetooButton
-          onClick={() => {}}
-          agree={agree}
-          agreeCount={agreeCount}
-          postId={postId}
+          handleClickMetooButton={handleClickMetooButton}
+          localMetooCount={localMetooCount}
         />
       </StButtonWrapper>
     </StWrapper>
